broke up reducers/state for orginization and readability; consolidated all useSelector calls into function for export and readability
