+++ frontend/src/App.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Route, BrowserRouter } from 'react-router-dom'
-import { useSelector } from 'react-redux';
+import { UI } from './redux/selectors';
+
 
 import Splash from './views/splash';
 import Tarot from './views/tarot';
@@ -11,7 +12,7 @@ const vh = window.innerHeight * .01;
 document.documentElement.style.setProperty('--vh', `${vh}px`)
 
 function App() {
-  const darkMode = useSelector((state) => state.setDarkMode.darkMode)
+  const { darkMode } = UI()
   const bgColor = darkMode ? { backgroundColor: '#212121' } : { backgroundColor: '#FAFAFA' }
   
   return (
