+++ frontend/src/features/UI/UIReducer.js
@@ -1,0 +1,13 @@
+import { combineReducers } from "redux";
+
+import composerReducer from "./composer/composerReducer";
+import darkModeReducer from './darkMode/darkModeSlice'
+import menuToggleReducer from './menu/menuSlider/menuSliderSlice'
+import modalsReducer from "./modals/modalsReducer";
+
+export default combineReducers({
+  composerReducer,
+  darkModeReducer,
+  menuToggleReducer,
+  modalsReducer
+})
