+++ frontend/src/features/UI/composer/color/color.scss
