+++ frontend/src/features/UI/composer/color/editor/editor.scss
