+++ frontend/src/features/UI/composer/color/editor/slider/rangeSlice.js
@@ -1,0 +1,15 @@
+import { createSlice } from "@reduxjs/toolkit";
+import initColor from '../../../../../../assets/static/init-color'
+
+export const rangeSlice = createSlice({
+  name: 'changeColor',
+  initialState: { colorData: initColor},
+  reducers: {
+    changeColor: (state, action) => {
+      state.colorData = action.payload
+    }
+  },
+})
+
+export const { changeColor } = rangeSlice.actions
+export default rangeSlice.reducer
