+++ frontend/src/features/UI/composer/color/editor/slider/slider.scss
