+++ frontend/src/features/UI/composer/composer.scss
