+++ frontend/src/features/UI/composer/editorSliderSlice.js
