+++ frontend/src/features/UI/menu/menu.scss
