+++ frontend/src/features/UI/menu/menuSlider/menu-slider.scss
