+++ frontend/src/features/UI/menu/menuSlider/menuSliderSlice.js
