+++ frontend/src/features/UI/modals/modalsReducer.js
@@ -1,0 +1,7 @@
+import { combineReducers } from "redux";
+
+import userModalReducer from './userModal/userModalSlice'
+
+export default combineReducers({
+  userModalReducer
+})
