+++ frontend/src/features/UI/modals/userModal/signUp/signUp.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import '../user-modal.scss'
+
+const SignUp = () => {
+  return (
+    <div>
+      <h2>Sign Up</h2>
+    </div>
+  )
+}
+
+export default SignUp
