+++ frontend/src/features/UI/top-bar/top-bar.scss
