--- frontend/src/features/modals/userModal/signUp.js
@@ -1,13 +1,0 @@
-import React from 'react'
-
-import './user-modal.scss'
-
-const SignUp = () => {
-  return (
-    <div>
-      <h2>Sign Up</h2>
-    </div>
-  )
-}
-
-export default SignUp
